Don't autoremove, as it causes container-not-found errors
Autoremove is equivalent to `docker run --rm`, and was causing a race
condition where we tried to check the exit status of a container after
it closed its stdout stream, but Docker was removing the container at
the same time.

This approach is more likely to leave containers lying around when the
user hits ctrl-C while running tc-builder, but those are easily cleaned
up with `docker container purge`.